Consolidate platform implementations for PDVideoPlayer
