+++ Sources/PDVideoPlayer/PDVideoPlayer.swift
@@ -1,30 +1,35 @@
-#if os(macOS)
 import SwiftUI
 import AVKit
 
 public struct PDVideoPlayerProxy<MenuContent: View> {
+#if os(macOS)
     public let player: PDVideoPlayerRepresentable<MenuContent>
+#else
+    public let player: PDVideoPlayerRepresentable
+#endif
     public let control: VideoPlayerControlView<MenuContent>
     public let navigation: VideoPlayerNavigationView
 }
 
 public struct PDVideoPlayer<MenuContent: View, Content: View>: View {
     @State private var model: PDPlayerModel? = nil
-    
+
     private var url: URL?
     private var player: AVPlayer?
-    
+
     var isMuted: Binding<Bool>?
     var playbackSpeed: Binding<PlaybackSpeed>?
     var onClose: VideoPlayerCloseAction?
     var onLongPress: VideoPlayerLongpressAction?
     var foregroundColor: Color = .white
+#if os(macOS)
     /// Enables moving the window when dragging on the player view.
     var windowDraggable: Bool = false
-    
+#endif
+
     private let content: (PDVideoPlayerProxy<MenuContent>) -> Content
     private let menuContent: () -> MenuContent
-    
+
     public init(
         url: URL,
         @ViewBuilder menu: @escaping () -> MenuContent,
@@ -34,7 +39,7 @@ public struct PDVideoPlayer<MenuContent: View, Content: View>: View {
         self.menuContent = menu
         self.content = content
     }
-    
+
     public init(
         player: AVPlayer,
         @ViewBuilder menu: @escaping () -> MenuContent,
@@ -44,14 +49,16 @@ public struct PDVideoPlayer<MenuContent: View, Content: View>: View {
         self.menuContent = menu
         self.content = content
     }
-    
+
     public var body: some View {
         if let model {
             let proxy = PDVideoPlayerProxy(
                 player: PDVideoPlayerRepresentable(
-                    model: model,
-                    playerViewConfigurator: { _ in },
+                    model: model
+#if os(macOS)
+                    , playerViewConfigurator: { _ in },
                     menuContent: menuContent
+#endif
                 ),
                 control: VideoPlayerControlView(
                     model: model,
@@ -59,7 +66,7 @@ public struct PDVideoPlayer<MenuContent: View, Content: View>: View {
                 ),
                 navigation: VideoPlayerNavigationView()
             )
-            
+
             content(proxy)
                 .videoPlayerKeyboardShortcuts(model)
                 .environment(model)
@@ -68,33 +75,31 @@ public struct PDVideoPlayer<MenuContent: View, Content: View>: View {
                 .environment(\.videoPlayerOnClose, onClose)
                 .environment(\.videoPlayerOnLongPress, onLongPress)
                 .environment(\.videoPlayerForegroundColor, foregroundColor)
-                .onChange(of: isMuted?.wrappedValue){
-                    if let isMuted{
+                .onChange(of: isMuted?.wrappedValue) { newValue in
+                    if let isMuted {
                         model.player.isMuted = isMuted.wrappedValue
                     }
                 }
-                .onChange(of: playbackSpeed?.wrappedValue){
+                .onChange(of: playbackSpeed?.wrappedValue) { _ in
                     if let speed = playbackSpeed?.wrappedValue {
                         model.playbackSpeed = speed
                     }
                 }
-                .onChange(of: url) {
-                    if let url {
-                        model.replacePlayer(url: url)
-                    }
+                .onChange(of: url) { url in
+                    if let url { model.replacePlayer(url: url) }
                 }
-                .onChange(of: player) {
-                    if let player{
-                        model.replacePlayer(with: player)
-                    }
+                .onChange(of: player) { player in
+                    if let player { model.replacePlayer(with: player) }
                 }
-        }else{
+        } else {
             Color.clear
-                .task{
+                .task {
                     if let url {
                         let m = PDPlayerModel(url: url)
                         m.onClose = onClose
+#if os(macOS)
                         m.windowDraggable = windowDraggable
+#endif
                         if let speed = playbackSpeed?.wrappedValue {
                             m.playbackSpeed = speed
                         }
@@ -102,7 +107,9 @@ public struct PDVideoPlayer<MenuContent: View, Content: View>: View {
                     } else if let player {
                         let m = PDPlayerModel(player: player)
                         m.onClose = onClose
+#if os(macOS)
                         m.windowDraggable = windowDraggable
+#endif
                         if let speed = playbackSpeed?.wrappedValue {
                             m.playbackSpeed = speed
                         }
@@ -112,4 +119,19 @@ public struct PDVideoPlayer<MenuContent: View, Content: View>: View {
         }
     }
 }
-#endif
+
+public extension PDVideoPlayer where MenuContent == EmptyView {
+    init(
+        url: URL,
+        @ViewBuilder content: @escaping (PDVideoPlayerProxy<MenuContent>) -> Content
+    ) {
+        self.init(url: url, menu: { EmptyView() }, content: content)
+    }
+
+    init(
+        player: AVPlayer,
+        @ViewBuilder content: @escaping (PDVideoPlayerProxy<MenuContent>) -> Content
+    ) {
+        self.init(player: player, menu: { EmptyView() }, content: content)
+    }
+}
